pyproject.toml
[project]
name = "rag-research-graph"
version = "0.0.1"
description = "Starter template for making a custom RAG research agent graph in LangGraph."
authors = [
    { name = "Vadym Barda", email = "19161700+vbarda@users.noreply.github.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=1.0.0,<2.0.0",
    "langchain-core>=1.2.0,<2.0.0",
    "langchain>=1.0.0,<2.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-fireworks>=1.0.0",
    "python-dotenv>=1.0.1",
    "langchain-postgres>=0.0.12",
    "psycopg[binary,pool]>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "langchain-pinecone>=0.1.3",
    "msgspec>=0.18.6",
    "langchain-mongodb>=0.1.9",
    "langchain-cohere>=0.5.0",
    "langsmith>=0.5.0",
    "langgraph-checkpoint>=3.0.1,<5.0.0",
    "protobuf>=6.32.1,<7.0.0",
    "click>=8.0.0",
    "langchain-community>=0.4.0",
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.3.0",
    "jinja2>=3.0.0",
    "pandas>=2.0.0",
    "edgartools>=1.0.0",
    "langchain-text-splitters>=0.2.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1", 
    "ruff>=0.6.1",
    "langgraph-cli",
    "langgraph-runtime-inmem"
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["index_graph", "shared", "index", "react_agent"]
[tool.setuptools.package-dir]
"langgraph.templates.index_graph" = "src/index_graph"
"index_graph" = "src/index_graph"
"shared" = "src/shared"
"index" = "src/index"
"react_agent" = "src/react_agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
[tool.pytest.ini_options]
pythonpath = [
  "src"
]
